Class level comments

+++ WorkoutTracker/src/main/java/com/example/workouttracker/repository/UserRepository.java
@@ -1,3 +1,12 @@
+/**
+ * UserRepository.java
+ * 
+ * Repository interface for user data access
+ * 
+ * Cody Crosby & Daniel Hanson
+ * CST-323: Cloud Computing
+ * GCU 2025
+ */
 package com.example.workouttracker.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +15,7 @@ import com.example.workouttracker.model.User;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
     Optional<User> findByEmail(String email);
 }
